gpu-worker/diarization/requirements.txt
# Diarization Worker Dependencies
# pyannote.audio for speaker diarization

self-hosted>=1.7.0
torch>=2.0.0
torchaudio>=2.0.0
soundfile>=0.12.0
numpy>=1.24.0
pyannote.audio>=3.3.0
